requirements.txt
chumpy
numpy==1.26.4
scipy
loguru
pytorch-lightning==2.2.0.post0
orjson
einops
c3d
pandas
BeautifulSoup4
matplotlib
configargparse
flatten_dict
yacs
smplx
git+https://github.com/nghorbani/human_body_prior.git
sk-video